requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.4
cffi==1.14.4
channels==3.0.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.0.11
django-autoslug==1.9.8
django-crequest==2018.5.11
django-environ==0.4.5
django-parler==2.2
djangorestframework==3.12.2
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
kombu==5.0.2
pep8==1.7.1
Pillow==8.0.1
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pyotp==2.4.1
pysha3==1.0.2
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
zope.interface==5.2.0
